backend/requirements.txt
# FastAPI 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 資料庫
motor==3.3.2
pymongo==4.6.0
# 移除 bson==0.5.10 因為它會與 pymongo 衝突

# 資料驗證
pydantic==2.5.0
pydantic-settings==2.1.0

# 認證和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
emails==0.6.0
jinja2==3.1.2

# 工具函數
python-dateutil==2.8.2
typing-extensions==4.8.0

# 開發工具
python-dotenv==1.0.0